Migrate and rename to NodeJS TestRunner
